# swagger-output.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ethers HD Wallet and Token Transfer Example",
    "version": "1.0.0",
    "description": "This REST API demonstrates how to use the Ethers library to interact with Ethereum using Infura. The API showcases various functionalities, including creating an HD wallet, fetching Ethereum and token balances, sending Ethereum and tokens, estimating gas prices, and more"
  },
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "local server"
    },
    {
      "url": "http://localhost:3000/",
      "description": "docker server"
    }
  ],
...
  "paths": {
...
    "/estimateGasPrice": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "get Ethereum gas price estimation",
        "description": "",
        "operationId": "estimateGasPrice",
        "responses": {
          "200": {
            "description": "Get gas price successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/gasPrice"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/gasPrice"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/getEthBalance": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Get logged-in user Ether balance",
        "description": "",
        "operationId": "getEthBalance",
        "responses": {
          "200": {
            "description": "Ether balance successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/etherBalance"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/etherBalance"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getTokenBalance": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "get logged-in user Token balance",
        "description": "",
        "operationId": "getTokenBalance",
        "responses": {
          "200": {
            "description": "Token balance successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tokenBalance"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/tokenBalance"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getAddrEthBalance/{address}": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "get specific address Ether balance",
        "description": "",
        "operationId": "getAddrEthBalance",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ethereum address"
          }
        ],
        "responses": {
          "200": {
            "description": "Ether balance successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/etherBalance"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/etherBalance"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getAddrTokenBalance/{address}": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "get specific address Token balance",
        "description": "",
        "operationId": "getAddrTokenBalance",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ethereum address"
          }
        ],
        "responses": {
          "200": {
            "description": "Token balance successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tokenBalance"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/tokenBalance"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getUserEthBalance/{id}": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "get specific user Ether balance",
        "description": "",
        "operationId": "getUserEthBalance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "User Ether balance successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ethBalance"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ethBalance"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/getUserTokenBalance/{id}": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "get specific user token balance",
        "description": "",
        "operationId": "getUserTokenBalance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "User token balance successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tokenBalance"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/tokenBalance"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/notFound"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/sendEthToAddr": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "send Ether from logged-in user to address",
        "description": "",
        "operationId": "sendEthToAddr",
        "parameters": [
          {
            "name": "address",
            "description": "Ethereum address",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Send Ether successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sendTransaction"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/sendTransaction"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToAddr"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/ToAddr"
              }
            }
          }
        }
      }
    },
    "/sendTokenToAddr": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "send Token from logged-in user to address",
        "description": "",
        "operationId": "sendTokenToAddr",
        "parameters": [
          {
            "name": "address",
            "description": "Ethereum address",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Send Token successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sendTransaction"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/sendTransaction"
                }
              }
            }
          },
          "500": {
            "description": "JWT Token error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToAddr"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/ToAddr"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "etherBalance": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "0xcdaa7e50203D9B48D3A181a02619c812Ce975CA4"
          },
          "balance": {
            "type": "string",
            "example": "0.00001"
          }
        },
        "xml": {
          "name": "etherBalance"
        }
      },
...
      "gasPrice": {
        "type": "object",
        "properties": {
          "gasPrice": {
            "type": "string",
            "example": "10.723000916 Gwei"
          },
          "maxFeePerGas": {
            "type": "string",
            "example": "22.44598146 Gwei"
          },
          "maxPriorityFeePerGas": {
            "type": "string",
            "example": "1.0 Gwei"
          }
        },
        "xml": {
          "name": "gasPrice"
        }
      },
...
      "notFound": {
        "type": "object",
        "properties": {
          "error": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Not found."
          },
          "stack": {
            "type": "string",
            "example": ""
          }
        },
        "xml": {
          "name": "notFound"
        }
      },
      "error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "number",
            "example": 500
          },
          "message": {
            "type": "string",
            "example": "Something went wrong"
          },
          "stack": {
            "type": "string",
            "example": ""
          }
        },
        "xml": {
          "name": "error"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}